DEV: Refactor measure_perf to scripts/utils.
